Moved interpro URL mapping and request processing into UploadController
