+++ src/main/java/sttDB/controller/UploadController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.multipart.MultipartHttpServletRequest;
 import sttDB.service.fastaServices.FastaParser;
+import sttDB.service.interproServices.InterproManager;
 
 import java.io.IOException;
 
@@ -17,9 +18,22 @@ public class UploadController {
     @Autowired
     FastaParser fastaParser;
 
+    private InterproManager interproManager;
+
     @PostMapping("/fasta")
     @ResponseBody
     public void recieveFasta(MultipartHttpServletRequest request) throws IOException{
         fastaParser.treatFasta(request);
+    }
+
+    @PostMapping("/interpro")
+    @ResponseBody
+    public void processRequest(MultipartHttpServletRequest request) {
+        interproManager.treatInterpro(request);
+    }
+
+    @Autowired
+    public void setInterproManager(InterproManager interproManager) {
+        this.interproManager = interproManager;
     }
 }
